requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
lxml
xmltodict
pycryptodome